+++ packages/localize-tools/src/tests/transform.unit.test.ts
@@ -51,6 +51,7 @@ function checkTransform(
   // Don't automatically load typings from nodes_modules/@types, we're not using
   // them here, so it's a waste of time.
   options.typeRoots = [];
+  options.experimentalDecorators = true;
   const result = compileTsFragment(inputTs, options, cache, (program) => ({
     before: [
       litLocalizeTransform(
@@ -444,13 +445,41 @@ test('different variable cast to "lit-localize-status" unchanged', (t) => {
   );
 });
 
-test('Localized(LitElement) -> LitElement', (t) => {
+test('updateWhenLocaleChanges -> undefined', (t) => {
   checkTransform(
     t,
     `import {LitElement, html} from 'lit';
-     import {Localized} from '@lit/localize/localized-element.js';
-     import {msg} from '@lit/localize';
-     class MyElement extends Localized(LitElement) {
+     import {msg, updateWhenLocaleChanges} from '@lit/localize';
+     class MyElement extends LitElement {
+       constructor() {
+         super();
+         updateWhenLocaleChanges(this);
+       }
+       render() {
+         return html\`<b>\${msg('Hello World!', {id: 'greeting'})}</b>\`;
+       }
+     }`,
+    `import {LitElement, html} from 'lit';
+     class MyElement extends LitElement {
+       constructor() {
+         super();
+         undefined;
+       }
+       render() {
+         return html\`<b>Hello World!</b>\`;
+       }
+     }`,
+    {autoImport: false}
+  );
+});
+
+test('@localized removed', (t) => {
+  checkTransform(
+    t,
+    `import {LitElement, html} from 'lit';
+     import {msg, localized} from '@lit/localize';
+     @localized()
+     class MyElement extends LitElement {
        render() {
          return html\`<b>\${msg('Hello World!', {id: 'greeting'})}</b>\`;
        }